// src/app/interfaces/archive.interface.ts
export interface IArchived {
    id: number;
    fileName: string;
    originalFileName: string;
    mimeType: string;
    fileSize: number;
    folderName: string;
    filePath: string;
    uploadDate: string;
    uploadedBy: string;
    tenantUniqueID: string;
    metadata: string;
    storageLocation: string;
    version: number;
    thumbnailPath: any;
  }
  